Separate pr samples into three TUs
