+++ tests/pr/backend/backend.vk.cc
@@ -24,7 +24,7 @@ TEST("pr::backend::vk sample_pbr", disabled, exclusive)
     // td::native::thread_sleep(15000);
     td::launch([&] {
         pr::backend::vk::BackendVulkan backend;
-        pr_test::run_sample(backend, get_vk_sample_conf());
+        pr_test::run_pbr_sample(backend, get_vk_sample_conf());
     });
 }
 
@@ -32,7 +32,7 @@ TEST("pr::backend::vk sample_cloth", disabled, exclusive)
 {
     td::launch([&] {
         pr::backend::vk::BackendVulkan backend;
-        pr_test::run_compute_sample(backend, get_vk_sample_conf());
+        pr_test::run_cloth_sample(backend, get_vk_sample_conf());
     });
 }
